fix png's with icc profile oddness
I'm sure some of you have seen the warnings generated by libpng 1.6:
"known incorrect sRGB profile". Now, this is no big deal, it's just a
warning, but I still wanted to look into it further. By my count,
there are 800 png files in Enlightenment, of which only 6 have
embedded iCC profiles, all from HP. The rest are sRGB colorspace, but
do not include a profile. There are some good arguments for this
(http://imageoptim.com/color-profiles.html http://www.gballard.net/psd/save_for_web_embed_ICC_profile.html)
and since this seems to be the default, I've attached the 6 files with
the profiles stripped out. As you can see, it doesn't change the image
or the colorspace, it just leaves the calibration up to the end user's
system.

I haven't been able to find just what's wrong with the HP iCC profile
being used that the authors of libpng don't like, but it makes sense
to standardize the way profiles are handled anyway. And it has the
nice side effect of silencing warnings, even if the Debian/Ubuntu
users aren't likely to see them for a couple of years.

I can just change the iCC profile instead if you would prefer. I've
seen multiple people recommend the Argyll sRGB profile, which is still
"sRGB IEC61966-2.1" and is public domain.