music-control: do not request props if the service is not up.
This prevent music player that support the dbus activation system
to be launched at module startup. Instead of blindly request properties
first check if the name has an owner.

As a side effect now the module fully support the dbus activation system:
if your player is not running it will be launched when you click his
name or any of the controls in the popup.

Note: If your player does not provide the autorun feature you can just
create a .service file in dbus share folder with the following content:

[D-BUS Service]
Name=org.mpris.MediaPlayer2.quodlibet
Exec=/usr/bin/quodlibet

...of course you must change the name of your player ;)