Entry: Fix update of child size and location.
Before this commit, setting the text of an entry after the entry
has already been placed and sized, would cause the inside entry
(the one inside of the scroller) size to be the size of the viewport
instead of the real size. This would make some parts of the entry
unreachable.

This fixes T169.