Entry: Fix scroller position to show cursor.
Before this commit, this would not always be the case.
I added a "scroll to cursor" for every change in entry size.

This is a further fix to T169. It's not exactly what described there,
but it's a similar issue I found while fixing it.

Somehow of a stage 2 for commit 80844cbcd72d5c3c00104b8bf8f9c5e590bad2bc.
This also adds changelog and news entries, as this issue is now fully fixed.