Cleanup PKGBUILD
Summary:
Name shouldn't include 17, that was only because it went into the Arch
repos that way
No need for _pkgver variable
Wayland is now at 1.3 in the repos, so the deps don't need to be
conditional. That was left over from before the release.
Removing libtool files is now the default, don't need it in the options.
Updated pkgver function. As it was, it would only include v_maj, v_min,
v_mic, but didn't take into account the -alpha/-beta/-rc added in the
AC_INIT. Now it does.
If we're building from git instead of snapshot tarballs as was
previously assumed, we need to use autogen.sh instead of just configure
Install text files to /usr/share/doc. This is more correct

Reviewers: cedric, raster

Reviewed By: raster

CC: cedric, raster

Differential Revision: https://phab.enlightenment.org/D380