clean up desklock showing on startup, add env var for checking unlocked state
desklock's layer is below the start splash layer, so this can just be shown any time regardless of whether it's enabled. security++

previously the fact that e was restarting could break out of desklocked startup due to how the env variables were set. now there is a new variable which only gets set when the screen is locked which will force lock on startup even during restarts

T681