add separate elm version requirement
elm versioning for stable series will be different than efl versioning, so these are not the same.

also now that the theme for enlightenment is distributed with elementary, there may be (and is) a MAXIMUM usable version of elementary which will work with a given E version while I work to stabilize various api across the releases. as a result, e18 RELEASES will not be usable with elm >= 1.9 (current git and beyond).

tl;dr: don't mix live versions with releases.