further reduce ibar menu hide delay, make delay even shorter when leaving menu
it seems that I always have this thing stuck open for a long period of time after moving my mouse out through it. I suspect that most times when a user moves the cursor out of the menu and does not immediately move it back or to the icon, the menu probably shouldn't hang around. likewise, when it moves out of the icon, maybe we give a little extra time to reach the menu, but not so much that it hangs around if the icon was exited through the side