e: call the right function without depending on ld to figure it out (mostly wrongly).
As a side note, using an EAPI function from within a module is a call for disaster.
Their is no guarantee from ld that it will be the current top open .so, actually it
is most likely not going to be the case. So it will do a search for the symbol in the
current stack of open .so and the first one to match will be chosen.

I do think it basically prove that current model of Enlightenment module is fairly
borken and error prone. Hope to have time to fix that, but it will involve completely
breaking the API and ABI compatibility of E module.