apply client resize limit for all possible size values during maximize calcs
Summary:
there was a problem that the app like gnome-terminal had not been restored to origin size after maximization.
I think that's because E overwrite the data of saved.x/y to maximized size in ConfigureRequest handler.
The reason why ConfigureReuqest happen is that the window tried to resize itself properly, since E resize window without regard of size hints.

Reviewers: zmike, seoz

CC: cedric

Differential Revision: https://phab.enlightenment.org/D633