bugfix: music_control: Use correct markup for metadata text.
We piped the plain metadata text to edje TEXTBLOCK which might interpret
some of the characters as markup. Better convert it to the correct markup
first. The actual bug case was a & in the metadata cut off the rest of
the text. Now we get the correct &amp before sending it of to edje.

Fixes T1061