i10n: Fix copy and paste bug in error translation. We fail SVG here not PNG.
This was really confusing. I always wondered why a message showed up in my E
startup that it was not able to load PNG files. It clearly could do so. Even
more confusing was that E did not shutdown immediately when detecting the
PNG problem. Turns out that that it was simply a wrong translation. SVG failed.