e backlight - limit min to 0.05 to avoid totally off backlight
this seems new, but if you set backlight to 0 it isnt at minimum, but
is entirely off on my intel laptop and this is not what was desired
with bl level of 0, so limit to 0.05 for now to avoid accidentally
turning your screen off. might fix this later to allow explicit off -
but this should use dpms etc. to turn the screen off as well as
possible scanout etc.