tiling: Fix issue with window-moving and windows not breaking out.
Summary:
The module is now working like it is described in T1773.

What the patch really does:
If a join request can not be done cause there is no node in the
direction to join, the join will try to break out the node into the
grand-grand-parent, this means a node can break out of his parent
without walking done each parent.

Issue as described in the original ticket:
Each number represents a different window:

122
134
155

Focus on 3, press Win+Left

I'd expect it to become:
132
134
135

But instead, nothing happens.

Fixes T1773

Reviewers: tasn

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D1632