modules/geolocation: Add communication to geoClue2 to receive current location.
This basically is the GeoClue2 client functionality a normal application also
uses. We request the current location and subscribe to updates on it.

Only activated when the users does a left click on the gadget so we do not
run always run it in the background. The received information will be displayed
in a popup in a following commit.