Tiling: Fixed break out method
Summary:
This fixes the case when par is NULL, this means we should insert a
parent node between the 2 nodes and root. (T1790)

Example:

Each Number is a Window,

12
Focus on 2

Press Win+Up should end up in this:

2
1

This patch also makes the bahavior for 2 windows more
consistent

12
Focus on 2

Pressing Win+Right,

21

It Will just swap the nodes,

This is the same behavior then in higher levels.

Reviewers: tasn

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D1643