e_win: Fix Dialog Titles
We should not set the title to "E" in the trap function of show.
If we do this we will override the icccm title and the netwm name, this
mean e_comp_x will set E as title of the window if the call to
elm_win_title_set was BEFORE the window was shown.

Now the name which was set before will be inserted.