Wizard: Update icon theme search list
Summary:
@fix The list of icon themes in the search list is quite out of date
it doesn't include the icon themes used by gnome3 or kde4/5 this means
for a lot of users at first boot e has alot of missing icons.

The Icon sets that have been added are as follows
      "Oxygen",            /* KDE 4 */
      "Adwaita",           /* Gnome 3 */
      "Breeze",            /* KDE 5 */
      "HighContrast"

This change will cause T1732 and T1923 to not occur for almost all users
but it does not fix them properly the code should be modified to fall back
to pick up icons from fallback locations such as the hicolor theme if no
icon theme is set.

It would be nice if someone could backport this to e19 I don't know how to

Reviewers: bu5hm4n, zmike, raster

Reviewed By: raster

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D2377