modules/geolocation: Follow InUse property change to give UI indication
The InUse property of the manager object is available in newer GeoClue2
versions. It shows if the service is in use by any application at the
moment. We use this to indicate the usage in the UI.
A blue geolocation icon means the service is in use, the grey icon means
it is inactive.