Changed search pattern to focus window by name to case-insensitive.
Summary: Sometimes a window has a name that's not like the application name, for example.  chromium-browser application will have a window name: Chromium-browser.  Most users will try to match the window name with "chromium-browser", but it wont work b/c the e's window name match is case sensitive.  Most users, would not know that the window name of chrome is "Chromium-browser", so it is pretty much impossible for them to setup a focus-to-window-name binding properly.

Reviewers: zmike

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D2645